fix: misc fixes
